Fix issues with Java 9+
In Java 9, the boot classpath system property was removed. There doesn't
seem to be a replacement. I can't find any way to duplicate what this
functionality was previously doing. The solution I came up with is to
use the Package/getPackages function, which at least gets the tests
passing and my basic functional tests working.

The `javax.xml.ws` package was removed in Java 11. See
https://blog.codefx.org/java/java-11-migration-guide/